First powershell version
This is a first version of a Powershell port of the timeroast attack script. It lacks some core functionality as the specification of custom RID ranges and the timeout parameter implementation.

For the former, the implementation is not as trivial as the Python one, since even if the default Powershell argument parser is really powerful, it does not support `num_ranges` variable types (AFAIK).

For the latter, I did not bother to add the timeout before trying to fix the async problem described below. 

Finally, this version outputs its results only when the execution is finished while the Python one is an async implementation, meaning that the results will be out-putted "on the go".

**Let op!** I am opening a merge request since GitHub is asking me to do so while creating the branch, feel free to close / ignore it.